Fix type error downstream
What's going on is that with the generic function signature before this
commit, `hashObj` was rejecting any object that didn't have the sign and
tag keys. I don't understand why this is happening because T was
supposed to extend this obj with empty keys. Anywhom, now it is straight
up structural typing with no generic.